# openapi.yaml
openapi: 3.1.0
info:
  title: ecommerce-backend
  description: |
    E-commerce backend API with authentication, cart, orders, payments, and admin operations.

    ## Authentication
    Most endpoints require a Bearer JWT token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Rate Limiting
    - `/auth/token`: 5 requests/minute
    - `/auth/refresh`: 10 requests/minute

    ## Idempotency
    Checkout and retry-payment endpoints require an `Idempotency-Key` header for safe retries.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /cart:
    get:
      tags:
        - cart
      summary: Get shopping cart
      description: Retrieve the current user's shopping cart
      operationId: getCart
      responses:
        "200":
          description: Cart retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartRead'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "404":
          $ref: '#/components/responses/NotFoundError'
    delete:
      tags:
        - cart
      summary: Clear cart
      description: Remove the current user's cart together with all of its items
      operationId: clearCart
      responses:
        "204":
          description: Cart cleared successfully
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "404":
          $ref: '#/components/responses/NotFoundError'
  /cart/items:
    post:
      tags:
        - cart
      summary: Add item to cart
      description: Add a new item to the shopping cart
      operationId: addCartItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItemCreate'
            examples:
              addItem:
                summary: Add item example
                value:
                  product_id: 1
                  quantity: 2
      responses:
        "201":
          description: Item added to cart successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartItemRead'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "409":
          $ref: '#/components/responses/ConflictError'
  /cart/items/{item_id}:
    patch:
      tags:
        - cart
      summary: Update cart item
      description: Update the quantity of an item in the cart
      operationId: updateCartItem
      parameters:
        - name: item_id
          in: path
          required: true
          description: Unique identifier of the cart item
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItemUpdate'
            examples:
              updateItem:
                summary: Update item example
                value:
                  quantity: 3
      responses:
        "200":
          description: Cart item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartItemRead'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "404":
          $ref: '#/components/responses/NotFoundError'
    delete:
      tags:
        - cart
      summary: Remove item from cart
      description: Remove an item from the shopping cart
      operationId: removeCartItem
      parameters:
        - name: item_id
          in: path
          required: true
          description: Unique identifier of the cart item
          schema:
            type: integer
            example: 1
      responses:
        "204":
          description: Item removed from cart successfully
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "404":
          $ref: '#/components/responses/NotFoundError'
  /cart/merge:
    post:
      tags:
        - cart
      summary: Merge cart items
      description: Merge multiple cart items into the user's cart
      operationId: mergeCart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CartItemCreate'
            examples:
              mergeItems:
                summary: Merge items example
                value:
                  - product_id: 1
                    quantity: 2
                  - product_id: 3
                    quantity: 1
      responses:
        "200":
          description: Cart merged successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartRead'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  schemas:
...
    CartItemCreate:
      type: object
      required:
        - product_id
      properties:
        product_id:
          type: integer
          minimum: 1
          description: Unique identifier of the product to add
          example: 1
        quantity:
          type: integer
          minimum: 1
          default: 1
          description: Quantity of the product to add
          example: 2
    CartItemUpdate:
      type: object
      required:
        - quantity
      properties:
        quantity:
          type: integer
          minimum: 1
          description: New quantity for the cart item
          example: 3
    CartItemRead:
      type: object
      required:
        - id
        - cart_id
        - product_id
        - quantity
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique cart item identifier
          example: 1
        cart_id:
          type: integer
          description: Identifier of the cart containing this item
          example: 1
        product_id:
          type: integer
          description: Identifier of the product
          example: 1
        quantity:
          type: integer
          description: Quantity of the product in the cart
          example: 2
        created_at:
          type: string
          format: date-time
          description: Timestamp when the item was added to cart
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the item was last updated
          example: "2024-01-20T14:25:00Z"
    CartRead:
      type: object
      required:
        - id
        - user_id
        - created_at
        - updated_at
        - items
      properties:
        id:
          type: integer
          description: Unique cart identifier
          example: 1
        user_id:
          type: integer
          description: Identifier of the user who owns the cart
          example: 1
        created_at:
          type: string
          format: date-time
          description: Timestamp when the cart was created
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the cart was last updated
          example: "2024-01-20T14:25:00Z"
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItemRead'
          description: List of items in the cart
          example: []
...
  responses:
    UnauthorizedError:
      description: Unauthorized - Invalid or missing authentication token
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: object
                required:
                  - type
                  - message
                properties:
                  type:
                    type: string
                    description: Error type/class name
                    example: "AuthenticationError"
                  message:
                    type: string
                    description: Human-readable error message
                    example: "Invalid or expired token."
          examples:
            missingToken:
              summary: Missing authorization header
              value:
                error:
                  type: "HTTPException"
                  message: "Missing authorization header."
            invalidToken:
              summary: Invalid token
              value:
                error:
                  type: "HTTPException"
                  message: "Invalid or expired token."
            invalidCredentials:
              summary: Invalid login credentials
              value:
                error:
                  type: "AuthenticationError"
                  message: "Invalid email or password."
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: object
                required:
                  - type
                  - message
                properties:
                  type:
                    type: string
                    description: Error type/class name
                    example: "NotFoundError"
                  message:
                    type: string
                    description: Human-readable error message
                    example: "User not found."
          examples:
            userNotFound:
              summary: User not found
              value:
                error:
                  type: "NotFoundError"
                  message: "User not found."
            orderNotFound:
              summary: Order not found
              value:
                error:
                  type: "NotFoundError"
                  message: "Order not found."
            productNotFound:
              summary: Product not found
              value:
                error:
                  type: "NotFoundError"
                  message: "Product not found."
...
    ConflictError:
      description: Conflict - Resource already exists or business rule violation
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: object
                required:
                  - type
                  - message
                properties:
                  type:
                    type: string
                    description: Error type/class name
                    example: "EmailAlreadyExistsError"
                  message:
                    type: string
                    description: Human-readable error message
                    example: "Email already exists."
          examples:
            emailExists:
              summary: Email already exists
              value:
                error:
                  type: "EmailAlreadyExistsError"
                  message: "Email already exists."
            cartItemExists:
              summary: Cart item already exists
              value:
                error:
                  type: "ValidationError"
                  message: "Cart item already exists."
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication.

        Obtain a token by calling `/auth/token` with email and password.

        Include the token in the Authorization header:
        ```
        Authorization: Bearer <access_token>
        ```

        The token contains the user ID in the `sub` claim and expires after 30 minutes by default.